+++ control/dcbf_optimizer.py
@@ -1,7 +1,8 @@
-from utils import *
 import numpy as np
 import casadi as ca
 import datetime
+
+from models.geometry_utils import *
 
 
 class NmpcDcbfOptimizerParam:
@@ -202,19 +203,3 @@ class NmpcDbcfOptimizer:
         delta_timer = end_timer - start_timer
         print("solver time: ", delta_timer.total_seconds())
         return opt_sol
-
-
-class NmpcDcbfController:
-    # TODO: Refactor this class to inheritate from a general optimizer
-    def __init__(self, dynamics=None):
-        self._param = NmpcDcbfOptimizerParam()
-        self._optimizer = NmpcDbcfOptimizer({}, {}, dynamics.forward_dynamics_opt(0.1))
-
-    def generate_control_input(self, system, global_path, local_trajectory, obstacles):
-        self._optimizer.setup(self._param, system, local_trajectory, obstacles)
-        self._opt_sol = self._optimizer.solve_nlp()
-        return self._opt_sol.value(self._optimizer.variables["u"][:, 0])
-
-    def logging(self, logger):
-        logger._xtrajs.append(self._opt_sol.value(self._optimizer.variables["x"]).T)
-        logger._utrajs.append(self._opt_sol.value(self._optimizer.variables["u"]).T)
